extracted reading and writing post functionality to own classes.
