+++ wpXml2Jekyll/wpXml2Jekyll.csproj
@@ -55,6 +55,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="PostImporter.cs" />
+    <Compile Include="PostWriter.cs" />
     <Compile Include="UIForm.cs">
       <SubType>Form</SubType>
     </Compile>
